replace es_size_t with size_t
Thanks to Pavel Levshin for the suggestion to use size_t in the new
API. And thanks to Pavel and David Lang in the following discussion
on the speed of 32 vs 64 bit variables in regard to the stack.

+++ src/liblognorm.h
@@ -212,6 +212,6 @@ int ln_loadSamples(ln_ctx ctx, char *file);
  *
  * @return Returns zero on success, something else otherwise.
  */
-int ln_normalize(ln_ctx ctx, const char *str, int strLen, struct json_object **json_p);
+int ln_normalize(ln_ctx ctx, const char *str, size_t strLen, struct json_object **json_p);
 
 #endif /* #ifndef LOGNORM_H_INCLUDED */
